fix rubocop

+++ app/models/course.rb
@@ -102,16 +102,14 @@ class Course < ApplicationRecord
     users_data = fetch_users_from_canvas(token, role)
 
     # Extract the Canvas user IDs from the fetched data
-    current_canvas_user_ids = users_data.map { |user_data| user_data['id'] }
+    current_canvas_user_ids = users_data.pluck('id')
 
     # Find all UserToCourse records for this course and role
     existing_user_to_courses = UserToCourse.where(course_id: id, role: role)
 
     # Delete UserToCourse records for users no longer in the course
     existing_user_to_courses.each do |user_to_course|
-      unless current_canvas_user_ids.include?(user_to_course.user.canvas_uid)
-        user_to_course.destroy
-      end
+      user_to_course.destroy unless current_canvas_user_ids.include?(user_to_course.user.canvas_uid)
     end
 
     # Create or update User and UserToCourse records for current users
